+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - 添加打开与复制菜单项 | Add open and copy menu items.
 - 添加刷新菜单项 | Add refresh menu item.
 - 添加新建文件/文件夹菜单项 | Add new file/folder menu item.
-- 添加文件/文件夹重命名菜单栏 | Add file/folder rename menu item.
+- 添加重命名文件/文件夹菜单项 | Add rename file/folder menu item.
+- 添加删除文件/文件夹菜单项 | Add delete file/folder menu item.
 
 ## 2023-07-23